requirements.txt
git+https://github.com/cleverhans-lab/cleverhans.git#egg=cleverhans
matplotlib
numpy
tqdm
absl-py
easydict
scikit-learn
torch==1.9.0
torchvision==0.10.0
skimage